Ellipsoid, Triangle, and translation
Comments have been translated into English.
A Triangle class has been added to work as the base element for upcoming
meshes. An Ellipsoid class has been added, and the Triangle class is
being used here for its mesh. Added a visibility function for the
Triangle's objects so that the system knows whether to draw them or not
when observing it from the observer's viewpoint at (0, 0, 0).